refactor: rename tt_table -> ttable
bench: 2204463
